Move SslErrorDialog into Application
This patch makes it possible to cut out a large piece from the lib. We
want to avoid having GUI code in the daemon. For Mirall, this is now
handled in Application. for other clients, they can do whatever they
want with this signal, and implement their own SSL Error handling.

The patch isn't exactly a beauty queen, since it involves some
back-and-forth between application and owncloudinfo objects, but it
seems the only way to properly separate and abstract the UI out of
owncloudinfo, and cuts down a huge part of the dependency chain,
especially QWidget-based classes.

I haven't been able to properly test this, but if anyone can give me a
login to a server that has wrong or untrusted SSL certificates, I'll
happily do it.

Merge? *puppy eyes*