Hacks to find INotify header on Debian
This patch forces INotify to be enabled, even if it's not found. It also
works around a bug in Debian's cmake which doesn't automatically set the
correct include path for the inotify.h header. Other distros will have
it in /usr/include/sys/inotify, but debian uses a multiarch approach
here.

Can be cleaned up once debian bug is fixed, test by just having it look
for sys/inotify.h.